+++ app/src/main/java/com/xpeho/xpeapp/MainActivity.kt
@@ -7,7 +7,7 @@ import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.material3.Surface
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.colorResource
-import com.xpeho.xpeapp.presentation.Home
+import com.xpeho.xpeapp.ui.presentation.Home
 import com.xpeho.xpeapp.ui.theme.XpeAppTheme
 
 class MainActivity : ComponentActivity() {