+++ app/src/main/java/com/xpeho/xpeapp/ui/presentation/componants/CustomDialog.kt
@@ -1,0 +1,24 @@
+package com.xpeho.xpeapp.ui.presentation.componants
+
+import androidx.compose.material3.*
+import androidx.compose.runtime.Composable
+
+@Composable
+fun ErrorDialog(
+    title: String,
+    message: String,
+    closeDialog: () -> Unit,
+) {
+    AlertDialog(
+        onDismissRequest = {},
+        title = { Text(text = title) },
+        text = { Text(message) },
+        confirmButton = {
+            Button(onClick = {
+                closeDialog()
+            }) {
+                Text("OK")
+            }
+        }
+    )
+}